• Update style for material table • Rename 'check status' button to 'search' • Sync update row and delete row to firebase and akita • Minor cleanup thoughout
